refactor: move component unit tests to a dedicated tests directory
