+++ tests/components/IconList.test.ts
@@ -1,9 +1,9 @@
 import { render, screen } from '@testing-library/svelte';
 import { describe, it, expect, vi, afterEach } from 'vitest';
-import IconList from './icons/IconList.svelte';
-import * as _icon from '../models/Icon';
+import IconList from '../../src/components/icons/IconList.svelte';
+import * as _icon from '../../src/models/Icon';
 
-const iconsPromise = vi.hoisted(() => import('../../tests/icon.store.mock'));
+const iconsPromise = vi.hoisted(() => import('../icon.store.mock'));
 
 
 // Mock the stores
