+++ tests/components/IconPreview.test.ts
@@ -1,10 +1,10 @@
-import { act, render, screen } from '@testing-library/svelte';
+import { render, screen } from '@testing-library/svelte';
 import { describe, it, expect, vi, afterEach } from 'vitest';
-import IconPreview from './icons/IconPreview.svelte';
-import { mkEmpty, type CustomizableIcon } from '../models/CustomizableIcon';
+import IconPreview from '../../src/components/icons/IconPreview.svelte';
+import { mkEmpty, type CustomizableIcon } from '../../src/models/CustomizableIcon';
 import { beforeEach } from 'node:test';
 
-const iconCustomizationsStorePromise = vi.hoisted(() => import('../../tests/icon-customizations.store.mock'));
+const iconCustomizationsStorePromise = vi.hoisted(() => import('../icon-customizations.store.mock'));
 
 vi.mock('../stores/icon-customizations.store', async () => await iconCustomizationsStorePromise);
